Added Mm and settings scene
they´re inside assets/development/Martín

as far as I know, the buttons work and lead to other scenes, and the volume settings work.

Lack a "back" button on settings, as I wait for it to be usable inside levels to learn how to make one.

May be a mess since no naming convention has been used yet