Added and changed fonts
Tried to not use more than 3 per scene, hard to tell which fits better for the game aesthetic as it is not defined yet. Put some more on Trello in case some seem a better fit.

Changed the layouts slightly.

Changes Scenes can be found in
Martin/Scenes/
Ending on V2